+++ src/GeoMath.cpp
@@ -190,25 +190,25 @@ bool testPolygonsForIntersect(const QVector<XY>& poly1, const QVector<XY>& poly2
 }
 
 // from http://www.movable-type.co.uk/scripts/LatLongVincenty.html
-float distance(const XY& p1, const XY& p2, float& a1, float& a2)
+double distance(const XY& p1, const XY& p2, double& a1, double& a2)
 {
-    float cosSigma = 0.0;
-    float sigma = 0.0;
-    float sinAlpha = 0.0;
-    float cosSqAlpha = 0.0;
-    float cos2SigmaM = 0.0;
-    float sinSigma = 0.0;
-    float sinLambda = 0.0;
-    float cosLambda = 0.0;
+    double cosSigma = 0.0;
+    double sigma = 0.0;
+    double sinAlpha = 0.0;
+    double cosSqAlpha = 0.0;
+    double cos2SigmaM = 0.0;
+    double sinSigma = 0.0;
+    double sinLambda = 0.0;
+    double cosLambda = 0.0;
 
-    float a = 6378137.0f, b = 6356752.3142f,  f = 1.0f/298.257223563f;  // WGS-84 ellipsiod
-    float L = p2.u - p1.u;
+    double a = 6378137.0f, b = 6356752.3142f,  f = 1.0f/298.257223563f;  // WGS-84 ellipsiod
+    double L = p2.u - p1.u;
 
-    float U1 = atan((1-f) * tan(p1.v));
-    float U2 = atan((1-f) * tan(p2.v));
-    float sinU1 = sin(U1), cosU1 = cos(U1);
-    float sinU2 = sin(U2), cosU2 = cos(U2);
-    float lambda = L, lambdaP = (float)(2*PI);
+    double U1 = atan((1-f) * tan(p1.v));
+    double U2 = atan((1-f) * tan(p2.v));
+    double sinU1 = sin(U1), cosU1 = cos(U1);
+    double sinU2 = sin(U2), cosU2 = cos(U2);
+    double lambda = L, lambdaP = (double)(2*PI);
     unsigned iterLimit = 20;
 
     while ((fabs(lambda - lambdaP) > 1e-12) && (--iterLimit > 0)) {
@@ -230,17 +230,17 @@ float distance(const XY& p1, const XY& p2, float& a1, float& a2)
             cos2SigmaM = 0;  // equatorial line: cosSqAlpha=0 (§6)
         }
 
-        float C = f/16 * cosSqAlpha * (4 + f * (4 - 3 * cosSqAlpha));
+        double C = f/16 * cosSqAlpha * (4 + f * (4 - 3 * cosSqAlpha));
         lambdaP = lambda;
         lambda = L + (1-C) * f * sinAlpha * (sigma + C*sinSigma*(cos2SigmaM + C * cosSigma * (-1 + 2 * cos2SigmaM * cos2SigmaM)));
     }
     if (iterLimit==0) return 0;  // formula failed to converge
 
-    float uSq = cosSqAlpha * (a*a - b*b) / (b*b);
-    float A = 1 + uSq/16384*(4096+uSq*(-768+uSq*(320-175*uSq)));
-    float B = uSq/1024 * (256+uSq*(-128+uSq*(74-47*uSq)));
-    float deltaSigma = B*sinSigma*(cos2SigmaM+B/4*(cosSigma*(-1+2*cos2SigmaM*cos2SigmaM)-B/6*cos2SigmaM*(-3+4*sinSigma*sinSigma)*(-3+4*cos2SigmaM*cos2SigmaM)));
-    float s = b*A*(sigma-deltaSigma);
+    double uSq = cosSqAlpha * (a*a - b*b) / (b*b);
+    double A = 1 + uSq/16384*(4096+uSq*(-768+uSq*(320-175*uSq)));
+    double B = uSq/1024 * (256+uSq*(-128+uSq*(74-47*uSq)));
+    double deltaSigma = B*sinSigma*(cos2SigmaM+B/4*(cosSigma*(-1+2*cos2SigmaM*cos2SigmaM)-B/6*cos2SigmaM*(-3+4*sinSigma*sinSigma)*(-3+4*cos2SigmaM*cos2SigmaM)));
+    double s = b*A*(sigma-deltaSigma);
 
     a1 = atan2(cosU2 * sinLambda, cosU1 * sinU2 - sinU1 * cosU2 * cosLambda) * 360 / TWOPI;
     a2 = atan2(cosU1 * sinLambda, -sinU1 * cosU2 + cosU1 * sinU2 * cosLambda) * 360 / TWOPI;
